Set version to 2.2.X for MACS2 on Python3. Set cython language level to 3str and edit pyx codes accordingly.

+++ MACS2/IO/Parser.pyx
@@ -1,4 +1,4 @@
-# Time-stamp: <2018-10-22 16:27:47 Tao Liu>
+# Time-stamp: <2018-10-24 13:33:01 Tao Liu>
 
 """Module for all MACS Parser classes for input.
 
@@ -180,7 +180,7 @@ cdef class GenericParser:
                 n += 1
         # done
         self.fhd.seek( 0 )
-        self.tag_size = s/n
+        self.tag_size = int(s/n)
         return self.tag_size
 
     cdef __tlen_parse_line ( self, bytes thisline ):
@@ -842,7 +842,7 @@ cdef class BAMParser( GenericParser ):
             s += a
             n += 1
         fseek( 0 )
-        self.tag_size = int( s/n )
+        self.tag_size = int(s/n)
         return self.tag_size
 
     cpdef tuple get_references( self ):
